First couple of posts
Both of these posts are poems. The first one, Detroit Annie
Hitchhiking, I used a `<div>` wrapper on each stanza. On the second
one, Self Evident, I used a class `.verse`, and created an `extra.css`
file to hold it. The CSS file is imported in `templates/blog-post.js`
so the CSS will be available in blog posts.

The Bio and layout both had hard-coded the starter kit's author, so
these were modified to use the info from `data/SiteConfig.js` instead.

I updated the `userAvatar` property to point to the actual gravatar image, and
added a `userPronouns` property as well.