Some addtions to try to work with React Components
It seems there's no easy way to grab frontmatter data and stuff it
into a Markdown file. The main suggestions seem to be to use React
components in the Markdown.

The `gatsby-remark-component` is one way to do this. Another
suggestion was to use
<https://github.com/ChristopherBiscardi/gatsby-mdx> which looks like a
huge undertaking right now.

I also wanted to figure out how to add tags and categories to posts,
and am looking at
<https://www.gatsbyjs.org/docs/adding-tags-and-categories-to-blog-posts/>
right now, and have added `gatsby-plugin-cateogies` and
`gatsby-plugin-tags`, but have not fully implemented them.

It seems that it's possible to add any sort of frontmatter key-value
pairs at will. It just requires a restart of the dev server.