Lots of work on getting Tags and Categories to work
The documentation from Gatsby v1 is an okay start, but it turns out to
be wrong in the end. I finally made this work by using the
`gatsby-plugin-tags` and `gatsby-plugin-categories` that take up the
heavy lifting. In the end, I only had to modify the template files to
make it work for my situation.